feat: transfer search into service
